@@ -1,15 +1,15 @@
 module github.com/cloudlena/adapters
 
-go 1.17
+go 1.18
 
 require (
-	github.com/golang-jwt/jwt/v4 v4.0.0
-	golang.org/x/oauth2 v0.0.0-20210810183815-faf39c7919d5
+	github.com/golang-jwt/jwt/v4 v4.3.0
+	golang.org/x/oauth2 v0.0.0-20220309155454-6242fa91716a
 )
 
 require (
 	github.com/golang/protobuf v1.4.2 // indirect
-	golang.org/x/net v0.0.0-20200822124328-c89045814202 // indirect
+	golang.org/x/net v0.0.0-20220127200216-cd36cc0744dd // indirect
 	google.golang.org/appengine v1.6.6 // indirect
 	google.golang.org/protobuf v1.25.0 // indirect
 )
